requirements.txt
python-dotenv
twitchio==2.7.0
